requirements.txt
flask
groq
langchain-community
langchain-pinecone
langchain-core
pinecone-client
sentence-transformers
python-dotenv
gunicorn